Revert change of UTs into ITs.
