+++ src/test/java/io/stargate/sgv2/jsonapi/service/cqldriver/CqlSessionCacheTest.java
@@ -29,8 +29,8 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 @QuarkusTest
-@TestProfile(CqlSessionCacheIntegrationTest.TestProfile.class)
-public class CqlSessionCacheIntegrationTest {
+@TestProfile(CqlSessionCacheTest.TestProfile.class)
+public class CqlSessionCacheTest {
   public static class TestProfile implements QuarkusTestProfile {
     // Alas, we do need actual DB backend so cannot do:
     // public boolean disableGlobalTestResources() { return true; }
